% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_brain_structs.R
\name{cifti_brain_structs}
\alias{cifti_brain_structs}
\title{Extract Brain Structures from CIFTI Brain Models}
\usage{
cifti_brain_structs(file)
}
\arguments{
\item{file}{\code{cifti} object}
}
\value{
A vector of brain structure names
}
\description{
Extracts the \code{BrainStructure} attribute from a
\code{BrainModel} in a \code{cifti} object
}
