% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_cifti.R
\name{cifti_coords_gifti}
\alias{cifti_coords_gifti}
\title{Map CIfTI values to Coordinates using GIfTI}
\usage{
cifti_coords_gifti(file, gii_file, structure, add_one = TRUE)
}
\arguments{
\item{file}{filename of CIFTI file or \code{cifti} object}

\item{gii_file}{filename of corresponding
GIFTI file or \code{gifti} object}

\item{structure}{Structure to map, must be one of the brain models in
the CIFTI}

\item{add_one}{should 1 be added to indices (1-based vs. 0-based)}
}
\value{
List of coordinates and values
}
\description{
Maps the data portion of CIFTI data set from a
Brain Model to the xyz coordinate triangles
}
