% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_data.R
\name{cifti_data}
\alias{cifti_data}
\title{Extract CIFTI data from file}
\usage{
cifti_data(fname, nim = NULL)
}
\arguments{
\item{fname}{Filename of CIFTI}

\item{nim}{NIfTI-2 header, if already parsed.
If \code{NULL}, \code{\link{nifti_2_hdr}}
will be run on the CIFTI.}
}
\value{
Array of values
}
\description{
Extracts the data after the CIFTI XML information
}
