% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read_char.R
\name{cifti_read_char}
\alias{cifti_read_char}
\title{Read characters with embedded nulls}
\usage{
cifti_read_char(fid, n, to = "UTF-8")
}
\arguments{
\item{fid}{identifier of the open file connection}

\item{n}{number of elements to read}

\item{to}{A character string describing the target encoding.}
}
\value{
Character vector
}
\description{
Simple wrapper for reading in character values with embedded nulls in a
binary file
}
