% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_xml.R
\name{cifti_xml}
\alias{cifti_xml}
\title{Extract CIFTI XML}
\usage{
cifti_xml(fname, nim = NULL)
}
\arguments{
\item{fname}{filename of CIFTI}

\item{nim}{NIfTI-2 header, if already parsed.
If \code{NULL}, \code{\link{nifti_2_hdr}}
will be run on the CIFTI.}
}
\value{
Character string of XML information
}
\description{
Extracts CIFTI XML from the CIFTI file
}

