% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cifti_data.R
\name{download_cifti_data}
\alias{download_cifti_data}
\title{Download CIFTI Test Data}
\usage{
download_cifti_data(outdir = system.file(package = "cifti"),
  overwrite = FALSE, ...)
}
\arguments{
\item{outdir}{Output directory for test file directory}

\item{overwrite}{Should files be overwritten if already exist?}

\item{...}{additional arguments to \code{\link{download.file}}}
}
\value{
Vector of file names
}
\description{
Downloads CIFTI test data from
\url{https://www.nitrc.org/frs/download.php/8541/cifti-2_test_data-1.2.zip}
}

