% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cifti_type.R
\name{get_cifti_type}
\alias{get_cifti_type}
\title{Get Generic CIFTI Type}
\usage{
get_cifti_type(fname, type = c("Volume", "Surface", "Parcel", "NamedMap",
  "BrainModel"))
}
\arguments{
\item{fname}{File name of CIFTI file}

\item{type}{type of data to extract from CIFTI XML}
}
\value{
List of output from each type
}
\description{
Wrapper for multiple types of CIFTI XML types.
}

