% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cifti_data.R
\name{have_cifti_test_data}
\alias{have_cifti_test_data}
\title{Check Presence of CIFTI Test Data}
\usage{
have_cifti_test_data(outdir = system.file(package = "cifti"))
}
\arguments{
\item{outdir}{Output directory for test file directory}
}
\value{
Logical indicator
}
\description{
Checks if CIFTI test data is downloaded
}

