% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti}
\alias{is.cifti}
\title{Test if CIFTI}
\usage{
is.cifti(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical if \code{x} is CIFTI
}
\description{
Simple wrapper to determine if class is CIFTI
}
