% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti_numeric.R
\name{is_cifti_numeric}
\alias{is_cifti_numeric}
\title{Test of numeric CIFTI field}
\usage{
is_cifti_numeric(x)
}
\arguments{
\item{x}{character vector of names}
}
\value{
Logical of length same as x
}
\description{
Wrapper of CIFTI fields for easy logical test
}

