% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_ind_map_nodes.R
\name{matrix_ind_map_nodes}
\alias{matrix_ind_map_nodes}
\title{Extract MatrixIndicesMap nodes from CIFTI}
\usage{
matrix_ind_map_nodes(fname)
}
\arguments{
\item{fname}{File of CIFTI data}
}
\value{
Nodes of class \code{xml_nodeset}
}
\description{
Extracts the nodes from a CIFTI-2 file corresponding to the
\code{MatrixIndicesMap} branch
}

