% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_brain_model.R
\name{parse_brain_model}
\alias{get_brain_model}
\alias{parse_brain_model}
\title{Parse BrainModel from CIFTI}
\usage{
parse_brain_model(nodeset)

get_brain_model(fname)
}
\arguments{
\item{nodeset}{Set of XML nodes corresponding to \code{BrainModel}}

\item{fname}{filename of CIFTI file}
}
\value{
List of values
}
\description{
Extracts information about BrainModels from CIFTI file
}
\examples{
\dontrun{
xmldata = cifti_xml(fname)
doc = read_xml(xmldata)
nodes = xml_find_all(doc, "/CIFTI/Matrix/MatrixIndicesMap")
nodeset = xml_find_all(nodes, "./BrainModel")
}
}

