% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_parcel.R
\name{parse_parcel}
\alias{parse_parcel}
\alias{get_parcel}
\title{Parse Parcel from CIFTI}
\usage{
parse_parcel(nodeset)

get_parcel(fname)
}
\arguments{
\item{nodeset}{Set of XML nodes corresponding to \code{Parcel}}

\item{fname}{filename of CIFTI file}
}
\value{
List of values
}
\description{
Extracts information about Parcels from CIFTI file
}
