% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surface.R
\name{parse_surface}
\alias{parse_surface}
\alias{get_surface}
\title{Parse Surface from CIFTI}
\usage{
parse_surface(nodeset)

get_surface(fname, verbose = TRUE)
}
\arguments{
\item{nodeset}{Set of XML nodes corresponding to \code{Surface}}

\item{fname}{filename of CIFTI file}

\item{verbose}{print diagnostic messages}
}
\value{
List of values
}
\description{
Extracts information about Surfaces from CIFTI file
}
\examples{
\dontrun{
doc = cifti_xml(fname)
nodes = xml_find_all(doc, "/CIFTI/Matrix/MatrixIndicesMap")
nodeset = xml_find_all(nodes, "./Surface")
parse_volume(nodeset)
}
}
