% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_volume.R
\name{parse_volume}
\alias{get_volume}
\alias{parse_volume}
\title{Parse Volume from CIFTI}
\usage{
parse_volume(nodeset)

get_volume(fname)
}
\arguments{
\item{nodeset}{Set of XML nodes corresponding to \code{Volume}}

\item{fname}{filename of CIFTI file}
}
\value{
List of values
}
\description{
Extracts information about Volumes from CIFTI file
}
\examples{
\dontrun{
xmldata = cifti_xml(fname)
doc = read_xml(xmldata)
nodes = xml_find_all(doc, "/CIFTI/Matrix/MatrixIndicesMap")
nodeset = xml_find_all(nodes, "./Volume")
parse_volume(nodeset)
}
}

